1.12-0.3.1 Adding jar file to the repository some fixes mcmod.info
Now compiled jar related to current tag will be in commit. Except
previous version tags because feature added just now.
Current jar is v0.3 but will be stored in this commit of version v0.3.1.
v0.3.1 it is just small changes in mcmod.info and it is also are not
directly related with not-source updating in repo.